requirements.txt
OpenAI
Flask
python-dotenv
chromadb
chromadb.utils
sentence-transformers
